refactor s3 upload into discrete module
