+++ src/actions/forecast/s3_upload.ts
@@ -1,0 +1,43 @@
+import { ActionRequest } from "../../hub"
+
+import * as S3 from "aws-sdk/clients/s3"
+import { PassThrough } from "stream"
+import * as winston from "winston"
+
+const striplines = require("striplines")
+
+export async function uploadToS3(request: ActionRequest, bucket: string, key: string) {
+  return new Promise<S3.ManagedUpload.SendData>((resolve, reject) => {
+    const s3 = new S3({
+      accessKeyId: request.params.accessKeyId,
+      secretAccessKey: request.params.secretAccessKey,
+    })
+
+    function uploadFromStream() {
+      winston.debug("calling uploadFromStream")
+      const passthrough = new PassThrough()
+
+      const params = {
+        Bucket: bucket,
+        Key: key,
+        Body: passthrough,
+      }
+      s3.upload(params, (err: Error|null, data: S3.ManagedUpload.SendData) => {
+        winston.debug("calling s3.upload")
+        if (err) {
+          return reject(err)
+        }
+        resolve(data)
+      })
+
+      return passthrough
+    }
+
+    request.stream(async (readable) => {
+      readable
+        .pipe(striplines(1))
+        .pipe(uploadFromStream())
+    }) // TODO: is this sensible error handle behavior?
+    .catch(winston.error)
+  })
+}
